only generate .fc when there's no static version

+++ src/builder/converter/cfgyaml.ts
@@ -42,23 +42,27 @@ export class CfgYamlConverter extends Converter {
           if (fs.existsSync(sourceCfgPath)) {
             const sourcePathWithoutExt = path.join(path.dirname(sourceCfgPath), path.basename(sourceCfgPath, '.cfg'));
             const targetPathWithoutExt = path.join(targetDirname, basename);
+            const variantPathWithoutExt = path.join(sourceDirname, basename);
 
-            // first try cf7
-            if (fs.existsSync(sourcePathWithoutExt + '.cf7')) {
-              cache.include(sourcePathWithoutExt + '.cf7');
-              try {
-                child.execFileSync(converterPath, ['-y', '-o', _dontOverwrite(targetPathWithoutExt + '.fc', '.fc'), '-w', sourcePathWithoutExt + '.cf7']);
-                this._logger.log(`  <= ${path.basename(targetPathWithoutExt)}.fc`);
+            // handle .fc/.cf7, but only if there's no variant specific file
+            if (!fs.existsSync(variantPathWithoutExt + '.cf7') && !fs.existsSync(variantPathWithoutExt + '.fc')) {
+              // first try cf7
+              if (fs.existsSync(sourcePathWithoutExt + '.cf7')) {
+                cache.include(sourcePathWithoutExt + '.cf7');
+                try {
+                  child.execFileSync(converterPath, ['-y', '-o', _dontOverwrite(targetPathWithoutExt + '.fc', '.fc'), '-w', sourcePathWithoutExt + '.cf7']);
+                  this._logger.log(`  <= ${path.basename(targetPathWithoutExt)}.fc`);
+                }
+                catch {
+                  this._logger.warn(`     AnnoFCConverter failed to write ${path.basename(targetPathWithoutExt)}.fc`);
+                }
               }
-              catch {
-                this._logger.warn(`     AnnoFCConverter failed to write ${path.basename(targetPathWithoutExt)}.fc`);
-              }
-            }
-            // then fc
-            else {
-              if (this._copyIfExists(sourcePathWithoutExt + '.fc', _dontOverwrite(targetPathWithoutExt + '.fc', '.fc'))) {
-                cache.include(sourcePathWithoutExt + '.fc', );
-                this._logger.log(`  <= ${path.basename(targetPathWithoutExt)}.fc`);
+              // then fc
+              else {
+                if (this._copyIfExists(sourcePathWithoutExt + '.fc', _dontOverwrite(targetPathWithoutExt + '.fc', '.fc'))) {
+                  cache.include(sourcePathWithoutExt + '.fc', );
+                  this._logger.log(`  <= ${path.basename(targetPathWithoutExt)}.fc`);
+                }
               }
             }
 
@@ -69,14 +73,30 @@ export class CfgYamlConverter extends Converter {
             fs.writeFileSync(_dontOverwrite(targetPathWithoutExt + '.cfg', '.cfg'), cfgContent.toString());
             this._logger.log(`  <= ${path.basename(targetPathWithoutExt)}.cfg`);
             
-            // read and modify ifo
-            if (fs.existsSync(sourcePathWithoutExt + '.ifo')) {
+            // read and modify ifo, but only if there's no variant specific ifo file
+            if (!fs.existsSync(variantPathWithoutExt + '.ifo') && fs.existsSync(sourcePathWithoutExt + '.ifo')) {
               cache.include(sourcePathWithoutExt + '.ifo');
               const ifoContent = AnnoXml.fromFile(sourcePathWithoutExt + '.ifo');
               this._runModifications(ifoContent, content.variant.ifo);
               fs.writeFileSync(_dontOverwrite(targetPathWithoutExt + '.ifo', '.ifo'), ifoContent.toString());
               this._logger.log(`  <= ${path.basename(targetPathWithoutExt)}.ifo`);
             }
+
+            // TODO
+            // let fcContent;
+            // // convert fc to cf7 if needed
+            // if (!fs.existsSync(sourcePathWithoutExt + '.cf7') && fs.existsSync(sourcePathWithoutExt + '.fc')) {
+            //   cache.include(sourcePathWithoutExt + '.fc');
+            // }
+            // // read and modify cf7
+            // if (!fcContent && !fs.existsSync(sourcePathWithoutExt + '.cf7')) {
+            //   fcContent = AnnoXml.fromFile(sourcePathWithoutExt + '.cf7');
+            // }
+            // if (fcContent) {
+            //   this._runModifications(fcContent, content.variant.fc);
+            //   fs.writeFileSync(_dontOverwrite(targetPathWithoutExt + '.fc', '.fc'), fcContent.toString());
+            //   this._logger.log(`  <= ${path.basename(targetPathWithoutExt)}.fc`);
+            // }
           }
           else {
             this._logger.error(`    ${sourceCfgPath} does not exist.`);
